Otel-log schema migrations, a couple bug fixes (#263)
1. Changed the schema migration path to send logs to otel-exporter if
available, taking care to flush the logger on exit.
2. Fixed a timing issue: otel-exporter's export() method used a callback
to notify when sent. Not waiting for the callback in a very short lived
command (like a failed migration) would cause telemetry data to never
get sent. Changed our calls to export to wait for the callback.
3. Increased the resolution of exported timestamps.
4. NIT: made logger.destroy() to call otel-collector.destroy() for just
a bit more syntax clarity.