Refactor DBOS Schema in System/User DB (#267)
This PR adds a `created_at` column in `dbos.transaction_outputs` table,
making it easier for garbage collection.

For garbage collection to work properly, this PR also refactors the way
we populate `created_at` and `updated_at` timestamps in the
workflow_status table. We record the timestamp when we first create the
internal workflow status object. The reason is that a workflow's
`created_at` be smaller than any of its transaction's `created_at`.
Otherwise, we might accidentally trim transaction outputs from other
workflows that are still actively used.

Previously, `created_at` for temporary workflows is set at the export
time, which is much later than the actual creation time (workflow status
buffer flushes every second, so the delay could be up to a second), and
would cause the above issue.