Foreign Key Constraints for System Tables (#273)
This PR adds foreign key constraints on DBOS system tables, making the
`workflow_staus.workflow_uuid` column as the core reference column for
all other tables, and configures cascading updates and deletes so we can
later correctly garbage collect these tables.

In addition, this PR changes behavior of the following operations:
- When we initialize workflow status, we now must synchronously set the
workflow's status to PENDING and record workflow inputs for all non
temp-transaction workflows. We have to do it for all types of workflows
because `operation_outputs` table has a foreign key constraint on
`workflow_status` table. We must ensure the workflow status exists
before we record any operation output.
- For temp-transaction workflows, we must buffer inputs after workflow
status is buffered (which is after the transaction completes) because
`workflow_inputs` table now has a foreign key reference to the
`workflow_status` table. Similarly, we also have to skip flushing a
workflow's inputs if its status is still in the buffer and not flushed
to the DB yet.
- The target workflow's status must exist when another function sends a
notification to it. Otherwise, inserting a non-exist `destinationUUID`
violates foreign key constraint on the `notifications` table, and we
throw a `DBOSNonExistWorkflowError`.

This PR updates several tests to comply with these changes.