Remove chmod (#284)
This is supposed to be OK:

1. Within package.json, things are executed by node and it knows what to
do without execute bit being set
2. `npm` will set the +x on things it installs that are listed in the
`bin` section. On Windows, its strategy is to create `.cmd` wrappers.

The exception is if someone wants to run cli.js directly from the shell
(which only works on Unix-like shells anyway). Then you need to change
the invocation to call `node`.