Open SDK version range for simple communicators to '*' (#292)
Also - missing files for nbgv (when run locally?) and touch up the
whitespace.

I think '*' is the fairest choice. Things like '>0.8.0' work for
everything except previews. Accepting `-preview` is a big pain because
it is limited to very specific previews. All these use from the SDK is
`@Communicator`, how bad can it be.