Global App Delete/Status Commands + Timestamp Issue (#317)
This PR adds an optional application name argument to `dbos-cloud app
status` and `delete` commands. This way, users can delete or view their
application status outside of an app folder, making it easier to install
the CLI globally and use it anywhere.

Another issue this PR fixes is the incorrect startTime in traces. We use
hrTime() to avoid the incorrect adjustment from opentelemetry.