Kafka Integration (#358)
Call a DBOS transaction or workflow upon receipt of a Kafka message. The
transaction or workflow is guaranteed to execute exactly-once for each
message received. Internally this is accomplished by setting the
idempotency key using the message's topic, partition, and offset, the
combination of which is guaranteed to be unique.

Example:

```javascript
import { KafkaConfig, KafkaMessage} from "kafkajs";

const kafkaConfig: KafkaConfig = {
    brokers: ['localhost:9092'] // For local testing
} 

@Kafka(kafkaConfig)
class KafkaExample{

  @KafkaConsume("example-topic") // You can optionally provide further consumer configuration
  @Workflow()
  static async exampleWorkflow(ctxt: WorkflowContext, topic: string, partition: number, message: KafkaMessage) {
    // This workflow executes exactly once for each message sent to the topic.
    // All methods annotated with Kafka decorators must take in the topic, partition, and message as inputs just like this method.
  }
}
```