Make CORS configurable / optional (#372)
I9 was trying to deploy a DBOS app with a react frontend. Calls w/
credentials to DBOS failed for CORS objections.

This PR makes CORS optional (so you could do your own CORS middleware;
as it sits the main CORS will prevent anything else from running), based
on the config .yaml file. It also allows credentials by default, if not
configured.

Includes a `@KoaCors` decorator that can override CORS policy for the
class. (Sort of like we are doing for body parser).

Once we have this PR settled, documentation will be needed. That is
being drafted now.