Workflow Invocation Methods (#432)
This PR introduces two new methods for invoking workflows from handlers.
For workflow bar in class Foo with argument arg:

- `ctxt.startWorkflow(Foo).bar(arg)` starts the workflow and returns its
handle. This is equivalent to the current `ctxt.invoke(Foo).bar(arg)`.
- `ctxt.invokeWorkflow(Foo).bar(arg)` executes the workflow and returns
its result. This is equivalent to the current
`ctxt.invoke(foo).bar(arg).then((x)=>x.getResult())`. The latter is very
unintuitive and is tripping users up.

To avoid confusion, we also deprecate calling `invoke` on workflows--use
one of the new methods instead.

My hope is that this makes it much clearer to users what is actually
happening when a handler invokes a workflow.