Revert "Reduce usage of "any" (#448)" (#469)
This reverts commit 8f1b5d4f44bb6dfb80a049693fb78f7125789451 as it
included breaking API changes. It caused numerous failures on our demo
applications and in our staging environment and cannot be released in
its current state. The changes in that PR are good, but we need to
introduce them gradually and carefully (likely in multiple smaller PRs)
and verify they don't cause regressions.