Allow proxy debug mode with undefined password (#485)
This PR fixes an issue where the undefined `PGPASSWORD` env in
`dbos-config.yaml` was parsed as `null` not an empty string. This makes
schema validation failed (i.e., debug mode with undefined password)
because `null` is not a `string` type. The fix is to move the check for
database password in `parseConfigFile` and assign the password to
`"PROXY-MODE"` if not set.

Tests:
- Added a test to make sure debug mode with a proxy allows undefined
`PGPASSWORD`.
- Fixed a minor issue in test where the proxy url wasn't correct and
caused the logger to print `: NaN` in tests.