Workflow Introspection (#520)
This PR introduces a new `getWorkflows()` method allowing users to query
the statuses of their workflows. For now, this is a method of
`HandlerContext`.

This method takes in the following input:

```typescript
export interface GetWorkflowsInput {
  workflowName?: string; // The name of the workflow function
  authenticatedUser?: string;
  startTime?: string; // In RFC 3339 format
  endTime?: string; // In RFC 3339 format
  status?: "PENDING" | "SUCCESS" | "ERROR";
  applicationVersion?: string;
}
```

It returns a list of the UUIDs of all workflows that satisfy all the
constraints specified in the input.

For example, to find the UUIDs of all executions of `testWorkflow` that
errored, run:

`getWorkflows({workflowName: "testWorkflow", status:
StatusString.Error})`

You can then call `retrieveWorkflow` on the retrieved UUID to obtain
more information about those specific workflows.