Fix Unintuitive Behavior (#573)
This fixes unintuitive behaviors that sparked user complaints:

1. Scheduled workflows now default to
`ExactlyOncePerIntervalWhenActive`. This means that if an executor is
started after being deactivated for a long time, it does not try to
"catch up" on past scheduled workflows. This is paired with a related
bugfix (https://github.com/dbos-inc/dbos-transact/pull/577) so this
behavior and flag now work as intended.

2. Communicator retries now clearly emit error messages that explain
what's going on, so they're no longer swallowed.
3. Fixed an issue where workflow errors were logged twice: once by the
workflow and once by the handler.