DBOSFailedSqlTransactionError (#640)
If a `@Transaction` method invokes a user-authored stored procedure that
raises an exception, the db transaction is aborted even if the exception
is caught in the `@Transaction` method. If the exception is caught and
handled in the `@transaction` method, DBOS will fail when it attempts to
write the `@Transaction` method return value to the
`transaction_outputs` table because the associated db transaction has
been aborted. This PR creates a DBOS specific error to throw in this
situation.

Fixes #626