do not start our http server during DBOS.launch() (#670)
Today this way of automatically starting our server is not compatible
with how we allow users to specify their own `start` command.

Specifically, if I pass a `start` command to `runtimeConfig`, the
current logic will fail and crash if there is already a server listening
on our default port (3000).

We can revisit what magic we want to do about that HTTP server, in the
future...