workflowID, template, and package doc links (#692)
Add 'workflowID' to WorkflowHandle, as getWorkflowUUID says UUID and
that's not strictly true any more.
Move "hello-v2" template to "hello". Old "hello" becomes "hello-context"
and is thusly marked at the top of the README.
Fix doc links (and a few bugs) in create templates, and library
packages.
Other minor touchups.