Prevent different workflows starting with the same ID (#726)
This PR addresses issue
https://github.com/dbos-inc/dbos-transact-ts/issues/725 where duplicate
workflow ID breaks the existing workflow. The solution is to check the
recorded function name, class name, and config name (if any) before
proceeding to start the workflow. If there's any difference, throw a
`DBOSConflictingWorkflowError` with a clear error message.

This PR also checks the recorded input versus the provided input, and
the provided queue name versus the original queue name, and print a
warning message if they don't match.

---------

Signed-off-by: Qian Li <liqian.cs@gmail.com>
Co-authored-by: Peter Kraft <peter.kraft@dbos.dev>
Co-authored-by: Peter Kraft <petereliaskraft@gmail.com>