V2 Stored Procedure Support (#697)
Adds `@DBOS.storedProcedure` (aka v2 stored procedures)

Note, plv8 does not support node's `AsyncLocalStorage` API. Since all
DBOS functions are static methods in order to support decorators, the
stored proc runtime uses
[`Function.apply`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Function/apply)
to provide a custom `this` object that the stored proc runtime version
of `DBOS` context methods accesses. This does change the expected
behavior of `this` object in static methods - normally, `this` would
point to the class object instance. However, access to the class object
instance via `this` isn't very useful to stored procedures anyway. If we
feel it's important to maintain access to the class object instance, we
can add a property to the custom `this` object that references the class
object.

---------

Co-authored-by: Chuck B <chuck.bear@dbos.dev>