Application Versioning (#765)
The PR expands the notion of application versions, which were previously
only used in DBOS Cloud, to support more robust self-hosting of DBOS
applications.

- When DBOS is launched, it constructs an application version by hashing
the source of the application's workflows. This version is stable across
restarts, but changes if the application's code is changed.
- The application version can be overridden through the
`DBOS__APPVERSION` environment variable.
- When recovering workflows, DBOS does not recover workflows belonging
to different application versions. This prevents code compatibility
issues.
- Workflows belonging to different app versions can be manually
recovered with `dbos workflow resume [workflow-ID]`.