do not rely on sleep (#770)
The previous version of this test was relying on a sleep to wait for a
task to be dequeued -- this proves flaky in GHA.

This PR modify the test to have:
1. Test worker send a notification each time a task is dequeued
2. Remove usage of sleep in the test
3. Wait for all the worker processes in a finally statement inside the
test workflow. This ensures that we don't fail the `afterEach` system db
shutdown, which in turn would fail the `beforeEach` statements of other
tests.