randomUUID  (#875)
update uuid package calls to node's built-in
[randomUUID](https://nodejs.org/docs/latest-v20.x/api/crypto.html#cryptorandomuuidoptions)
function

Note, we were using v4 UUIDs in the production code base but used v1 and
v4 UUIDs in the test code. As far as I can tell, there was no dependency
on the different semantics of v1 and v4 UUIDs in our test code, so I
replaced all of them with `randomUUID` (which returns a v4 UUID). Also,
there were a couple places in the test code where we validate UUIDs. But
that uuid validate function is just a regex call so I added a uuid
validate function to the test `helpers.ts` file