Update SysDB.listWorkflows / listQueuedWorkflows (#864)
This PR updates the implementation of several System DB methods related
to getting Workflow Status as well as several of the types related to
Workflow Status.

* System DB `getWorkflows` and `getQueuedWorkflows` have been updated to
match DBOS Python's implementation + now return full status info instead
of just a list of WF IDs.
* This simplifies and should improve performance for places like
Conductor `dispatchLoop` and workflow management CLI which originally
had to make separate DB round trips to retrieve full workflow status and
inputs for WF IDs returned by `listWorkflows` or `listQueuedWorkflows`
* System DB's `getWorkflowStatus` now returns `WorkflowStatusInternal`
and `getWorkflowStatusInternal` has been removed.
* `WorkflowStatusInternal` adds an optional `input` field for methods
like `getWorkflows` and `getQueuedWorkflows` that retrieve inputs.
* `WorkflowStatusInternal.maxRetries` has been removed from the
interface and added as an optional parameter to `initWorkflowStatus`.
* `WorkflowStatus` has been updated to have matching fields from
`WorkflowStatusInternal` but with deserialized input/output/error fields
* `WorkflowInformation` interface has been removed. Places that used it
have been updated to use `WorkflowStatus`

For discussion:
* Note, this PR originally did not change public facing APIs such as
[`DBOS.getWorkflows`](https://docs.dbos.dev/typescript/reference/transactapi/dbos-class#dbosgetworkflows).
However, it was decided that it was worth breaking the public facing
APIs in this case was worth it