Handle Database Disconnections (#930)
If a workflow encounters a database connection issue while performing an
operation, block the workflow and retry the operation until it
reconnects and succeeds.

In other words, if DBOS loses its database connection, everything pauses
until the connection is recovered, trading off availability for
correctness.

Add a suite of new "chaos tests" that validate DBOS can successfully
execute workflows in the presence of random database disconnects and
restarts.