use Proxy/Reflect in startWorkflow (#957)
`startWorkflow` creates a proxy object from an existing class or class
instance where the workflow methods return a WorkflowHandle. The
previous proxy creation approach was dependent on the workflow's
registered name being the same as the method name. With the new
`registerWorkflow` API, the developer can choose whatever name they
wish. Unless they pick the method name, `startWorkflow` no longer works.

This PR modifies `startWorkflow` to use JS's native
[Proxy](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Proxy)
object. The proxy handler's `get` method first uses
`getRegistrationForFunction` to retrieve the `MethodRegistration` for
the workflow function. If this fails, `startWorkflow` falls back to
matching on the workflow function name, as it did previously. This
enables `startWorkflow` to work with both decorated and registered
workflows, regardless of what the developer names the registered
workflow.