Change what happens to step/tx outside wf (#958)
Steps and transactions should run without workflows.

Generally, "calling step/tx outside of workflows will simply run the
bare functions without checkpointing into system tables." This also
means no WF status record also. A function need not have serializable
inputs unless it is a workflow.

The exception to this is if the user seems to be doing workflow-like
stuff, like asking for an assigned WFID. If they've asked for an
assigned ID, but the next function is not a workflow, we are giving an
error.

This is a significant part of the change, however some additional cases
would be added/changed when we clean out the context-based
implementation. That is, we should run the steps under the common
codepath, but without the WF context... making that happen is a bit
risky and will conflict with removal of the WF context. But, with this
change, the API is working. When we consolidate the code, the tracing
and step retry will be worth testing.

This also fixes:
- package.json exports for some usage scenarios
- A couple of `startWorkflow` cases that were previously working with
`startWorkflowFunction`
- POST param handling not matching v1/v2