use knex user db client in chaos test (#1018)
there appears to be an issue with a pooled pg-node user db throwing when
the underlying db goes away. The config PR (#1008) changed the default
from knex to pg-node, which caused the chaos test to fail. This PR
changes the chaos test back to using knex. I also filed #1017 to track
the `PGNodeUserDatabase` issue