Optional User DB (#1016)
Makes the user DB optional.

~If you're using `DBOS.setConfig` there is now a `enableUserDatabase`
field that *MUST* be set to true in order to have a user database (i.e.
it defaults to false).~ If you're using a config file instead of
`setConfig`, we enable a user database if the database.app_db_client
field is set.

> Update: We decided against adding the `enableUserDatabase` config
field. Instead, if `userDatabaseClient` is undefined, DBOS will not
create a user database at all.

SInce we are trying to move people off of the config file, I did not
want to add an enable user db field to it. However, this will break
users that have not set the app_db_client, as it previously defaulted to
Knex.