assert sysdb name is not empty + save orignal error on rethrow (#1033)
`PostgresSystemDatabase` retrieves the database name from the url in two
places. This PR adds an assert to ensure the db name is not empty.
Postgres's error in this scenario is confusing (`zero-length delimited
identifier at or near """"
`)

Other changes:
* `DBOSExecutor.init` catches errors during initialization and throws a
`DBOSInitializationError` in the original error's place...which loses
all the context of the original error. This PR adds an optional error
field to `DBOSInitializationError` to store the original error.
* Realized while making this change that we don't validate the db name
in `getSystemDatabaseUrl`, so I fixed that + added tests as well