Migration Permissions (#1042)
When creating the DBOS system tables with `npx dbos schema` you can now
specify an "application role."
If you specify this role, it is granted all permissions needed to run a
DBOS app.

```
Usage: dbos schema [options] [systemDatabaseUrl]

Create the DBOS system database and its internal tables

Arguments:
  systemDatabaseUrl        System database URL

Options:
  -d, --appDir <string>    Specify the application root directory
  -r, --app-role <string>  The role with which you will run your DBOS application
  -h, --help               display help for command
```

This is most useful in settings where the application runs with an
"application role" that does not have sufficient privileges to create
the DBOS system database or system tables. Prior to deployment, the dbos
migrate command can be run with a privileged user to create DBOS system
tables, specifying the application role to which to grant permissions on
those tables. Then, the application can safely run with the application
role.