Support `Promise.allSettled` (#1051)
Generically, make sure the function ID is incremented and then snapshot
all or part of the DBOS context prior to the first `await`, allowing
steps and step-like internal functions to proceed in parallel without
confusion.

Note that support of `Promise.all` is not a goal. `Promise.all` is
dangerous in node.js because of how it handles rejections.
`Promise.allSettled` is safer and users should be steered in that
direction.

---------

Co-authored-by: Harry Pierson <harry.pierson@dbos.dev>