Serializability checking (#1069)
When and how you find out about serializability issues, in reverse order
of desirability:
1. During recovery time, when you use the deserialized item instead of
the original and the code throws a weird error.
2. With any kind of error that consistently occurs at runtime during
testing
3. With a nice error that consistently occurs during testing
4. At compilation time
5. In the IDE / real time
6. It doesn't happen because you or your AI just know what you're doing

Main?  Currently at a 1.  This PR currently brings it to 3.

Compare the passing test a few commits back:

https://github.com/dbos-inc/dbos-transact-ts/blob/8f4a504fbe864a4da5d14c8b8391fdf0edc3eada/tests/decorator-free.test.ts#L619

To the current situation for that same file:

https://github.com/dbos-inc/dbos-transact-ts/blob/a960b2e16eda2760b5c0db5a6bc19fa632058195/tests/decorator-free.test.ts#L621