Single Workflow Execution (#1120)
Make it so that only one execution of a workflow is outstanding at a
time:
1. Do not start PENDING workflows, unless this is a dequeue or recovery
attempt. (Join w/ existing execution).
2. If a workflow is running locally already (say, a repeated recovery
attempt), do not run it. (Join w/ existing execution).
3. Change system DB step checkpoint behavior to report an error only if
the record was written by someone else, not if it exists (this can
happen in a DB connection hiccup scenario).

Tests for the above.

Also:
- Fix an issue w/ client-enqueued `send`
- Update some lib versions (secvulns)